Implement new logs UI
